DOUBLE -> FLOAT

+++ core/Param.cc
@@ -2,7 +2,7 @@
 
 Param::Param() : type(INT) { value.i = 0; }
 Param::Param(const int i) : type(INT) { value.i = i; }
-Param::Param(const double d) : type(DOUBLE) { value.d = d; }
+Param::Param(const double d) : type(FLOAT) { value.d = d; }
 
 Param::operator int() const {
 	if (type != INT)
@@ -11,7 +11,7 @@ Param::operator int() const {
 }
 
 Param::operator double() const {
-	if (type != DOUBLE)
+	if (type != FLOAT)
 		throw (new TypeMismatchError())->with(_POS);
 	return value.d;
 }
